refactor: @ Move findClosestParentTask() to ListItem.ts
Note that whilst the tests pass, the Tasks plugin does not load.
It crashes at startup, giving this console error:

Plugin failure: obsidian-tasks-plugin TypeError: Class extends value undefined is not a constructor or null
